Add "legacy" formula
The ReferenceFreeFormula class is supposed to imitate the original
Dice syntax and forbid the advanced reference elements.
Ideally this will just be a drop-in, but we'll see!

Also don't stress out about how to get the Rules from everywhere... they
don't seem to have widespread use so it's probably better to just use
the default rather than twisting our arm to inject them in weird places.