Play around with search efficiency
It turns out "compressing" the regex's representation does not make it
notably faster. In fact, the increased complexity seems to be vaguely
slower! That said, experimentation indicates that switching the groups
to be NON-capture brings the result up to about the same speed as the
original while saving a few characters.

Not really impressive. Perhaps directly executing the Trie on the string
would be faster than invoking regexes at all, but that's an experiment
for another day.